+++ src/util/featureflags.ts
@@ -1,1 +1,5 @@
-export const useConditionInFocus = true;
+import { getConfiguration } from "./vsccommon";
+
+const featureFlags = getConfiguration().featureFlags;
+
+export const useConditionInFocus = !featureFlags.includes('!useConditionInFocus');